Polish Developer Hub modal and ADB removal
