--- .github/workflows/pr150-followup.yml
@@ -1,51 +1,0 @@
-name: Apply PR 150 follow-up
-
-on:
-  push:
-    branches:
-      - agent/devhub-vrhotspot-only-wireless
-    paths:
-      - .github/workflows/pr150-followup.yml
-
-permissions:
-  contents: write
-
-jobs:
-  apply:
-    runs-on: ubuntu-24.04
-    steps:
-      - name: Check out PR branch
-        uses: actions/checkout@v4
-        with:
-          ref: agent/devhub-vrhotspot-only-wireless
-          fetch-depth: 0
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.12'
-
-      - name: Set up Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: '22.23.1'
-
-      - name: Apply guarded follow-up patch
-        run: python tools/ci/apply_pr150_followup.py
-
-      - name: Install validation dependencies
-        run: python -m pip install -e . ruff
-
-      - name: Validate production changes
-        run: |
-          python -m compileall -q backend/vr_hotspotd
-          ruff check backend/vr_hotspotd/devtools/platform_tools_manager.py backend/vr_hotspotd/devtools/devhub_api.py
-          node --check assets/devhub_connection_wizard.js
-
-      - name: Commit follow-up to PR branch
-        run: |
-          git config user.name "josethevrtech"
-          git config user.email "250434127+josethevrtech@users.noreply.github.com"
-          git add -A
-          git commit -m "Polish Developer Hub modal and ADB removal"
-          git push origin HEAD:agent/devhub-vrhotspot-only-wireless
